+++ scripts/test/ProcessManagement/TransitionAction/TicketCreate.t
@@ -609,6 +609,26 @@ my @Tests = (
         Success => 1,
         Article => 0,
     },
+    {
+        Name   => 'Correct Ticket->DynamicField_Field3_Value No Article',
+        Config => {
+            UserID => $UserID,
+            Ticket => \%Ticket,
+            Config => {
+                Title         => 'ProcessManagement::TransitionAction::TicketCreate::5::' . $RandomID,
+                CustomerID    => '123465',
+                CustomerUser  => 'customer@example.com',
+                OwnerID       => '1',
+                TypeID        => 1,
+                ResponsibleID => 1,
+                PendingTime   => '2014-12-23 23:05:00',
+
+                "DynamicField_Field1$RandomID" => "<OTRS_TICKET_DynamicField_Field3$RandomID" . '_Value>',
+            },
+        },
+        Success => 1,
+        Article => 0,
+    },
 
     {
         Name   => 'Correct Ticket->NotExistent',
@@ -789,7 +809,7 @@ for my $Test (@Tests) {
                 )
             {
                 $ExpectedValue = $Ticket{$1} // '';
-                if ( !ref $ExpectedValue ) {
+                if ( !ref $ExpectedValue && $OrigTest->{Config}->{Config}->{$Attribute} !~ m{_Value} ) {
                     $Self->IsNot(
                         $Test->{Config}->{Config}->{$Attribute},
                         $OrigTest->{Config}->{Config}->{$Attribute},
@@ -801,6 +821,28 @@ for my $Test (@Tests) {
                         "$ModuleName - Test:'$Test->{Name}' | Attribute: $Attribute value:",
                     );
 
+                }
+                elsif ( $OrigTest->{Config}->{Config}->{$Attribute} =~ m{OTRS_TICKET_DynamicField_(\S+?)_Value} ) {
+                    $Self->IsNot(
+                        $Test->{Config}->{Config}->{$Attribute},
+                        $OrigTest->{Config}->{Config}->{$Attribute},
+                        "$ModuleName - Test:'$Test->{Name}' | Attribute: $Attribute value: $OrigTest->{Config}->{Config}->{$Attribute} should been replaced",
+                    );
+                    my $DynamicFieldName = $1;
+
+                    my $DynamicFieldConfig = $DynamicFieldObject->DynamicFieldGet(
+                        Name => $DynamicFieldName,
+                    );
+                    my $DisplayValueStrg = $DynamicFieldBackendObject->ReadableValueRender(
+                        DynamicFieldConfig => $DynamicFieldConfig,
+                        Value              => $Ticket{"DynamicField_$DynamicFieldName"},
+                    );
+
+                    $Self->Is(
+                        $Test->{Config}->{Config}->{$Attribute},
+                        $DisplayValueStrg->{Value},
+                        "$ModuleName - Test:'$Test->{Name}' | Attribute: $Attribute value:",
+                    );
                 }
                 else {
                     $Self->IsNotDeeply(
@@ -831,10 +873,29 @@ for my $Test (@Tests) {
             # }
 
             if ( $Test->{Article} ) {
-                if ( !ref $ExpectedValue ) {
+                if ( !ref $ExpectedValue && $OrigTest->{Config}->{Config}->{$Attribute} !~ m{_Value} ) {
                     $Self->Is(
                         $Article{$ArticleAttribute},
                         $ExpectedValue,
+                        "$ModuleName - Test:'$Test->{Name}' | Attribute: $Attribute for ArticleID:"
+                            . " $Article{ArticleID} match expected value",
+                    );
+                }
+                elsif ( $OrigTest->{Config}->{Config}->{$Attribute} =~ m{OTRS_TICKET_DynamicField_(\S+?)_Value} ) {
+
+                    my $DynamicFieldName = $1;
+
+                    my $DynamicFieldConfig = $DynamicFieldObject->DynamicFieldGet(
+                        Name => $DynamicFieldName,
+                    );
+                    my $DisplayValueStrg = $DynamicFieldBackendObject->ReadableValueRender(
+                        DynamicFieldConfig => $DynamicFieldConfig,
+                        Value              => $Ticket{"DynamicField_$DynamicFieldName"},
+                    );
+
+                    $Self->Is(
+                        $Article{$ArticleAttribute},
+                        $DisplayValueStrg->{Value},
                         "$ModuleName - Test:'$Test->{Name}' | Attribute: $Attribute for ArticleID:"
                             . " $Article{ArticleID} match expected value",
                     );